transportation exceptions added

+++ case-backend/src/main/java/com/example/thy/service/TransportationService.java
@@ -2,11 +2,17 @@ package com.example.thy.service;
 
 import com.example.thy.dto.TransportationDto;
 import com.example.thy.entity.Transportation;
+import com.example.thy.exception.LocationAlreadyExistsException;
+import com.example.thy.exception.LocationNotFoundException;
+import com.example.thy.exception.TransportationAlreadyExistsException;
+import com.example.thy.exception.TransportationNotFoundException;
 import com.example.thy.repository.TransportationRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -24,13 +30,29 @@ public class TransportationService {
     }
 
     public TransportationDto findById(Long id) {
-        return modelMapper.map(transportationRepository.findById(id).orElse(null), TransportationDto.class);
+        Optional<Transportation> optional = transportationRepository.findById(id);
+        if (optional.isPresent()) {
+            return modelMapper.map(optional.get(), TransportationDto.class);
+        }else {
+            throw new TransportationNotFoundException("Transportation not found for find",id);
+        }
     }
     public TransportationDto save(TransportationDto transportationDto) {
         Transportation transportation = modelMapper.map(transportationDto, Transportation.class);
-        return modelMapper.map(transportationRepository.save(transportation), TransportationDto.class);
+        try{
+            transportation = transportationRepository.save(transportation);
+        }catch (DataIntegrityViolationException e){
+            log.error(e.getMessage());
+            throw new TransportationAlreadyExistsException(e.getMessage());
+        }
+        return modelMapper.map(transportation, TransportationDto.class);
     }
+    @Transactional
     public void deleteById(Long id) {
-        transportationRepository.deleteById(id);
+        if(transportationRepository.existsById(id)){
+            transportationRepository.deleteById(id);
+        }else{
+            throw new TransportationNotFoundException("Transportation not found for delete",id);
+        }
     }
 }
